added 'map' as valid datatype
in Databricks there is datatype 'map' for columns in a UnityCatalog table. After I created a datacontract, which includes the UnityCatalog table as "model", the linting of the datacontract fails....since 'map' is not in the list of valid types. Since I want to keep the original datatype, it would be great to have 'map' added.